+++ parser/src/states/cohorts_states/any/input_state.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Default)]
 pub struct InputState {
-    pub count: f32,
-    pub volume: f32,
+    pub count: f64,
+    pub volume: f64,
 }
 
 impl InputState {
-    pub fn iterate(&mut self, count: f32, volume: f32) {
+    pub fn iterate(&mut self, count: f64, volume: f64) {
         self.count += count;
         self.volume += volume;
     }
